Fix presentation export formatting, CAD 2D DXF parser, and modularize artifact compiler/importer
