+++ .gitignore
@@ -13,3 +13,4 @@ frontend/node_modules/
 frontend/dist/
 .DS_Store
 debug_*.json
+samples/
