+++ backend/artifacts/compiler/pptx.py
@@ -1,0 +1,891 @@
+"""
+backend/artifacts/compiler/pptx.py
+Compiles presentation artifacts into 100% compliant 16:9 Microsoft PowerPoint (.pptx)
+compatible with Apple Keynote, Google Slides, and MS Office.
+
+Implements the DOM-to-PowerPoint Engine: translates runtime generative HTML5/CSS slides,
+Bento cards, metric callouts, images, and speaker notes directly into native DrawingML shapes.
+"""
+import io
+import os
+import re
+import json
+import shutil
+import tempfile
+import subprocess
+from typing import List, Any, Dict, Optional
+
+import lxml.html
+from pptx import Presentation
+from pptx.util import Inches, Pt
+from pptx.dml.color import RGBColor
+from pptx.enum.shapes import MSO_SHAPE
+from pptx.enum.text import PP_ALIGN
+from pptx.oxml import parse_xml
+
+from models import SessionArtifact
+from .utils import (
+    _strip_html,
+    _sanitize_xml_text,
+    _parse_hex_color,
+    _fetch_image_bytes,
+)
+
+
+def _embed_pptx_image(slide, img_bytes: bytes, box_left, box_top, box_w, box_h):
+    """
+    Safely embed image bytes with proportional aspect ratio into slide without distortion.
+    """
+    from PIL import Image
+    try:
+        with Image.open(io.BytesIO(img_bytes)) as pil_img:
+            img_w, img_h = pil_img.size
+        if img_w <= 0 or img_h <= 0:
+            return None
+        img_ar = img_w / img_h
+        box_ar = box_w / box_h
+
+        if img_ar > box_ar:
+            draw_w = box_w
+            draw_h = int(box_w / img_ar)
+            draw_left = int(box_left)
+            draw_top = int(box_top + (box_h - draw_h) / 2)
+        else:
+            draw_h = box_h
+            draw_w = int(box_h * img_ar)
+            draw_top = int(box_top)
+            draw_left = int(box_left + (box_w - draw_w) / 2)
+
+        pic = slide.shapes.add_picture(
+            io.BytesIO(img_bytes),
+            draw_left,
+            draw_top,
+            draw_w,
+            draw_h
+        )
+        return pic
+    except Exception:
+        return None
+
+
+def _compile_generative_slide_dom(
+    slide,
+    slide_html: str,
+    default_title: str = "Slide Title",
+    notes_text: str = ""
+):
+    """
+    Translates a single generative HTML5 slide into native PowerPoint DrawingML shapes and textboxes.
+    """
+    clean_html = slide_html.strip() if slide_html else ""
+    if not clean_html:
+        clean_html = f"<div><h1>{default_title}</h1></div>"
+
+    try:
+        doc = lxml.html.fragment_fromstring(clean_html, create_parent='div')
+    except Exception:
+        # Fallback if unparseable HTML
+        doc = lxml.html.fragment_fromstring(f"<div><h1>{default_title}</h1><p>{_strip_html(clean_html)}</p></div>", create_parent='div')
+
+    # 1. Background Color Extraction
+    root_style = doc.get("style", "")
+    bg_color = _parse_hex_color(root_style, RGBColor(9, 13, 22))
+    fill = slide.background.fill
+    fill.solid()
+    fill.fore_color.rgb = bg_color
+
+    # Determine contrast theme
+    luminance = 0.299 * bg_color[0] + 0.587 * bg_color[1] + 0.114 * bg_color[2]
+    is_light = luminance > 160
+    default_text_color = RGBColor(15, 23, 42) if is_light else RGBColor(255, 255, 255)
+    default_subtext_color = RGBColor(71, 85, 105) if is_light else RGBColor(148, 163, 184)
+    default_card_bg = RGBColor(241, 245, 249) if is_light else RGBColor(21, 28, 44)
+    default_accent = RGBColor(37, 99, 235) if is_light else RGBColor(129, 140, 248)
+
+    # 2. Extract Badge / Pill (e.g. <div class="badge">KEYNOTE</div>)
+    badge_matches = doc.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " badge ") or contains(concat(" ", normalize-space(@class), " "), " pill ")]')
+    badge_el = badge_matches[0] if badge_matches else None
+    badge_text = _sanitize_xml_text(badge_el.text_content().strip()) if badge_el is not None else ""
+    badge_color = _parse_hex_color(badge_el.get("style", "") if badge_el is not None else "", default_accent)
+
+    # 3. Extract Slide Main Title
+    h_matches = doc.xpath('.//h1 | .//h2')
+    h1_el = h_matches[0] if h_matches else None
+    title_text = _sanitize_xml_text(h1_el.text_content().strip()) if h1_el is not None else default_title
+    title_color = _parse_hex_color(h1_el.get("style", "") if h1_el is not None else "", default_text_color)
+
+    # 4. Extract Subtitle
+    sub_matches = doc.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " subtitle ") or contains(concat(" ", normalize-space(@class), " "), " desc ")]')
+    sub_el = sub_matches[0] if sub_matches else None
+    if sub_el is None:
+        all_p = doc.xpath('.//p')
+        if all_p:
+            sub_el = all_p[0]
+    subtitle_text = _sanitize_xml_text(sub_el.text_content().strip()) if sub_el is not None else ""
+    if subtitle_text == title_text:
+        subtitle_text = ""
+
+    # ── Render Slide Header ──
+    current_y = 0.6
+    if badge_text:
+        badge_box = slide.shapes.add_shape(
+            MSO_SHAPE.ROUNDED_RECTANGLE,
+            Inches(0.8),
+            Inches(current_y),
+            Inches(min(max(len(badge_text) * 0.15, 1.8), 3.5)),
+            Inches(0.35)
+        )
+        badge_box.fill.solid()
+        badge_box.fill.fore_color.rgb = RGBColor(30, 27, 75) if not is_light else RGBColor(238, 242, 255)
+        badge_box.line.color.rgb = badge_color
+        badge_box.line.width = Pt(1)
+        tf = badge_box.text_frame
+        tf.word_wrap = True
+        tf.margin_left = Inches(0.05)
+        tf.margin_right = Inches(0.05)
+        tf.margin_top = Inches(0.05)
+        tf.margin_bottom = Inches(0.05)
+        p = tf.paragraphs[0]
+        p.text = badge_text.upper()
+        p.font.size = Pt(10)
+        p.font.bold = True
+        p.font.color.rgb = badge_color
+        p.alignment = PP_ALIGN.CENTER
+        current_y += 0.45
+
+    title_box = slide.shapes.add_textbox(Inches(0.8), Inches(current_y), Inches(11.7), Inches(0.9))
+    tf = title_box.text_frame
+    tf.word_wrap = True
+    tf.margin_left = Inches(0)
+    tf.margin_top = Inches(0)
+    p = tf.paragraphs[0]
+    p.text = title_text
+    p.font.size = Pt(32)
+    p.font.bold = True
+    p.font.color.rgb = title_color
+    current_y += 0.95
+
+    if subtitle_text:
+        sub_box = slide.shapes.add_textbox(Inches(0.8), Inches(current_y), Inches(11.7), Inches(0.6))
+        tf = sub_box.text_frame
+        tf.word_wrap = True
+        tf.margin_left = Inches(0)
+        tf.margin_top = Inches(0)
+        p = tf.paragraphs[0]
+        p.text = subtitle_text
+        p.font.size = Pt(14)
+        p.font.color.rgb = default_subtext_color
+        current_y += 0.65
+
+    # ── Check for Standalone Image (e.g. Hero banner or Split column) ──
+    img_matches = doc.xpath('.//img')
+    img_src = img_matches[0].get("src") if img_matches else None
+
+    # ── 5. Extract Containers / Cards (Bento Grids, Feature Boxes, Metrics) ──
+    all_cards = doc.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " card ") or contains(concat(" ", normalize-space(@class), " "), " box ") or contains(concat(" ", normalize-space(@class), " "), " pillar ") or contains(concat(" ", normalize-space(@class), " "), " column ")]')
+
+    def is_nested(el, pool):
+        for other in pool:
+            if other is not el and other in el.iterancestors():
+                return True
+        return False
+
+    top_cards = [c for c in all_cards if not is_nested(c, all_cards)]
+
+    if top_cards:
+        card_count = len(top_cards)
+
+        # A) Split Layout: Image alongside cards
+        if img_src and card_count <= 2:
+            card_y = max(current_y + 0.1, 2.5)
+            avail_h = 7.0 - card_y
+            left_w = 5.8
+            c_h = (avail_h - (0.2 * (card_count - 1))) / card_count
+
+            for idx, card in enumerate(top_cards):
+                c_top = card_y + idx * (c_h + 0.2)
+                c_style = card.get("style", "")
+                c_bg = _parse_hex_color(c_style, default_card_bg)
+                c_border = _parse_hex_color(re.search(r'border(?:-color)?:\s*[^;]*', c_style).group(0) if re.search(r'border(?:-color)?:\s*[^;]*', c_style) else "", default_accent)
+
+                shape = slide.shapes.add_shape(MSO_SHAPE.ROUNDED_RECTANGLE, Inches(0.8), Inches(c_top), Inches(left_w), Inches(c_h))
+                shape.fill.solid()
+                shape.fill.fore_color.rgb = c_bg
+                shape.line.color.rgb = c_border
+                shape.line.width = Pt(1)
+
+                tf = shape.text_frame
+                tf.word_wrap = True
+                tf.margin_left = Inches(0.2)
+                tf.margin_top = Inches(0.2)
+                c_title_matches = card.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " title ") or contains(concat(" ", normalize-space(@class), " "), " heading ") or @class="card-title"]')
+                c_title = c_title_matches[0] if c_title_matches else None
+                p_first = tf.paragraphs[0]
+                if c_title is not None:
+                    p_first.text = _sanitize_xml_text(c_title.text_content().strip())
+                    p_first.font.size = Pt(16)
+                    p_first.font.bold = True
+                    p_first.font.color.rgb = default_text_color
+                    p_first.space_after = Pt(6)
+
+                for p_el in card.xpath('.//p | .//li'):
+                    p_text = _sanitize_xml_text(p_el.text_content().strip())
+                    if p_text and (c_title is None or p_text != _sanitize_xml_text(c_title.text_content().strip())):
+                        p_b = tf.add_paragraph()
+                        p_b.text = f"• {p_text}" if p_el.tag == 'li' else p_text
+                        p_b.font.size = Pt(12)
+                        p_b.font.color.rgb = default_subtext_color
+                        p_b.space_after = Pt(4)
+
+            # Embed image in right column
+            img_bytes = _fetch_image_bytes(img_src)
+            img_placed = False
+            if img_bytes:
+                pic = _embed_pptx_image(slide, img_bytes, Inches(7.0), Inches(card_y), Inches(5.5), Inches(avail_h))
+                if pic is not None:
+                    img_placed = True
+
+            if not img_placed:
+                # Placeholder frame
+                img_shape = slide.shapes.add_shape(MSO_SHAPE.ROUNDED_RECTANGLE, Inches(7.0), Inches(card_y), Inches(5.5), Inches(avail_h))
+                img_shape.fill.solid()
+                img_shape.fill.fore_color.rgb = default_card_bg
+                img_shape.line.color.rgb = default_accent
+                img_shape.line.width = Pt(1)
+                tf = img_shape.text_frame
+                tf.word_wrap = True
+                p = tf.paragraphs[0]
+                p.text = "[Visual Image]"
+                p.font.size = Pt(13)
+                p.font.color.rgb = default_accent
+                p.alignment = PP_ALIGN.CENTER
+
+        # B) Bento Grid Layout: 1, 2, 3, 4, 5, 6 cards
+        else:
+            max_cols = 3 if card_count in (3, 5, 6) else (2 if card_count == 2 else min(card_count, 4))
+            rows = 1 if card_count <= max_cols else 2
+
+            card_y = max(current_y + 0.1, 2.5)
+            avail_h = 7.0 - card_y
+            card_h = (avail_h - (0.25 * (rows - 1))) / rows
+            total_w = 11.733
+            gap = 0.25
+            card_w = (total_w - (gap * (max_cols - 1))) / max_cols
+
+            for idx, card in enumerate(top_cards):
+                r = idx // max_cols
+                c = idx % max_cols
+                c_left = 0.8 + c * (card_w + gap)
+                c_top = card_y + r * (card_h + 0.25)
+
+                c_style = card.get("style", "")
+                c_bg = _parse_hex_color(c_style, default_card_bg)
+                c_border = _parse_hex_color(re.search(r'border(?:-color)?:\s*[^;]*', c_style).group(0) if re.search(r'border(?:-color)?:\s*[^;]*', c_style) else "", default_accent)
+
+                shape = slide.shapes.add_shape(MSO_SHAPE.ROUNDED_RECTANGLE, Inches(c_left), Inches(c_top), Inches(card_w), Inches(card_h))
+                shape.fill.solid()
+                shape.fill.fore_color.rgb = c_bg
+                shape.line.color.rgb = c_border
+                shape.line.width = Pt(1)
+
+                stat_matches = card.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " stat ") or contains(concat(" ", normalize-space(@class), " "), " metric ") or contains(concat(" ", normalize-space(@class), " "), " number ")]')
+                stat = stat_matches[0] if stat_matches else None
+                c_title_matches = card.xpath('.//*[contains(concat(" ", normalize-space(@class), " "), " title ") or contains(concat(" ", normalize-space(@class), " "), " heading ") or @class="card-title"]')
+                c_title = c_title_matches[0] if c_title_matches else None
+                c_ps = card.xpath('.//p | .//li')
+
+                tf = shape.text_frame
+                tf.word_wrap = True
+                tf.margin_left = Inches(0.2)
+                tf.margin_right = Inches(0.2)
+                tf.margin_top = Inches(0.2)
+                tf.margin_bottom = Inches(0.2)
+
+                p_first = tf.paragraphs[0]
+                if stat is not None:
+                    p_first.text = _sanitize_xml_text(stat.text_content().strip())
+                    p_first.font.size = Pt(28)
+                    p_first.font.bold = True
+                    stat_col = _parse_hex_color(stat.get("style", ""), c_border)
+                    p_first.font.color.rgb = stat_col
+                    p_first.space_after = Pt(4)
+
+                    if c_title is not None:
+                        p_t = tf.add_paragraph()
+                        p_t.text = _sanitize_xml_text(c_title.text_content().strip())
+                        p_t.font.size = Pt(14)
+                        p_t.font.bold = True
+                        p_t.font.color.rgb = default_text_color
+                        p_t.space_after = Pt(6)
+                elif c_title is not None:
+                    p_first.text = _sanitize_xml_text(c_title.text_content().strip())
+                    p_first.font.size = Pt(16)
+                    p_first.font.bold = True
+                    p_first.font.color.rgb = default_text_color
+                    p_first.space_after = Pt(6)
+                else:
+                    p_first.text = ""
+
+                for p_el in c_ps:
+                    p_text = _sanitize_xml_text(p_el.text_content().strip())
+                    if p_text and (c_title is None or p_text != _sanitize_xml_text(c_title.text_content().strip())):
+                        p_body = tf.add_paragraph()
+                        p_body.text = f"• {p_text}" if p_el.tag == 'li' else p_text
+                        p_body.font.size = Pt(11.5)
+                        p_body.font.color.rgb = default_subtext_color
+                        p_body.space_after = Pt(4)
+
+    else:
+        # C) Fallback: Bullet list or paragraphs without card containers
+        all_items = doc.xpath('.//li | .//p')
+        # Exclude subtitle if already displayed
+        display_items = [it for it in all_items if _sanitize_xml_text(it.text_content().strip()) != subtitle_text and _sanitize_xml_text(it.text_content().strip()) != title_text]
+
+        if display_items:
+            body_box = slide.shapes.add_textbox(Inches(0.8), Inches(current_y + 0.2), Inches(11.7), Inches(4.2))
+            tf = body_box.text_frame
+            tf.word_wrap = True
+            tf.margin_left = Inches(0)
+            tf.margin_top = Inches(0)
+
+            for i, it in enumerate(display_items):
+                p = tf.paragraphs[0] if i == 0 else tf.add_paragraph()
+                text_clean = _sanitize_xml_text(it.text_content().strip())
+                if it.tag == 'li' or text_clean.startswith(('-', '*', '•')):
+                    p.text = f"• {text_clean.lstrip('-*• ')}"
+                else:
+                    p.text = text_clean
+                p.font.size = Pt(15)
+                p.font.color.rgb = default_subtext_color
+                p.space_after = Pt(10)
+
+    # 6. Speaker Notes
+    if notes_text and notes_text.strip():
+        slide.notes_slide.notes_text_frame.text = _sanitize_xml_text(notes_text)
+
+
+def _ensure_notes_master_registered(prs: Presentation):
+    """
+    Ensures that if any slide has speaker notes, the notesMaster relationship is properly
+    registered in presentation.xml with <p:notesMasterIdLst>.
+    Apple Keynote strictly enforces ECMA-376 schema rules: if a notesMaster relationship
+    exists in presentation.xml.rels, it must be declared in <p:notesMasterIdLst> inside
+    presentation.xml, otherwise Keynote rejects the file as 'invalid format'.
+    """
+    notes_master_rid = None
+    for rId, rel in prs.part.rels.items():
+        if "notesMaster" in rel.reltype:
+            notes_master_rid = rId
+            break
+
+    if notes_master_rid and not prs._element.xpath("//p:notesMasterIdLst"):
+        xml_str = (
+            f'<p:notesMasterIdLst xmlns:p="http://schemas.openxmlformats.org/presentationml/2006/main" '
+            f'xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships">'
+            f'<p:notesMasterId r:id="{notes_master_rid}"/>'
+            f'</p:notesMasterIdLst>'
+        )
+        elem = parse_xml(xml_str)
+        sldMasterIdLst = prs._element.xpath("//p:sldMasterIdLst")
+        if sldMasterIdLst:
+            idx = prs._element.index(sldMasterIdLst[0])
+            prs._element.insert(idx + 1, elem)
+        else:
+            prs._element.insert(0, elem)
+
+
+def _sanitize_slide_html(html: str) -> str:
+    """Cleans up escaped quotes, entities, and malformed SVG attributes."""
+    if not html:
+        return ""
+    h = html.replace("&quot;", '"').replace(r'\"', '"').replace(r"\'", "'")
+    h = h.replace('""', '"')
+    h = re.sub(r'viewBox="0"[^>]*?fill=', 'viewBox="0 0 24 24" fill=', h)
+    return h
+
+
+NODE_DOM_EXTRACTION_SCRIPT = """
+const puppeteer = require("/usr/lib/node_modules/puppeteer-core");
+const fs = require("fs");
+
+(async () => {
+  const input = JSON.parse(fs.readFileSync(0, "utf-8"));
+  const browser = await puppeteer.launch({
+    executablePath: "/usr/bin/chromium",
+    args: ["--no-sandbox", "--disable-gpu", "--disable-dev-shm-usage"]
+  });
+
+  const results = [];
+  for (const item of input) {
+    let cleanHtml = item.html;
+    const page = await browser.newPage();
+    await page.setViewport({ width: 1280, height: 720 });
+    await page.setContent(`<!DOCTYPE html><html><head><meta charset="utf-8"><style>* { margin: 0; padding: 0; box-sizing: border-box; } html, body { width: 1280px; height: 720px; overflow: hidden; font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif; }</style></head><body>${cleanHtml}</body></html>`);
+
+    const data = await page.evaluate(() => {
+      function parseRgba(str) {
+        if (!str) return null;
+        const m = str.match(/rgba?\\s*\\(\\s*(\\d+)\\s*,\\s*(\\d+)\\s*,\\s*(\\d+)(?:\\s*,\\s*([\\d.]+))?\\)/);
+        if (!m) return null;
+        return {
+          r: parseInt(m[1], 10),
+          g: parseInt(m[2], 10),
+          b: parseInt(m[3], 10),
+          a: m[4] !== undefined ? parseFloat(m[4]) : 1.0
+        };
+      }
+
+      function parseGradientOrColor(style) {
+        const bg = style.background || style.backgroundColor || "";
+        const m = bg.match(/#(?:[0-9a-fA-F]{6}|[0-9a-fA-F]{3})\\b/);
+        if (m) return m[0];
+        const rgb = parseRgba(style.backgroundColor);
+        if (rgb && rgb.a > 0.1) return `rgb(${rgb.r}, ${rgb.g}, ${rgb.b})`;
+        return "#0f172a";
+      }
+
+      const body = document.body;
+      const rootDiv = body.firstElementChild || body;
+      const rootStyle = window.getComputedStyle(rootDiv);
+
+      function isInsideSvg(el) {
+        return el.closest("svg") !== null;
+      }
+
+      // 1. SVGs
+      const svgs = [];
+      const svgEls = Array.from(document.querySelectorAll("svg"));
+      svgEls.forEach((svg, idx) => {
+        const rect = svg.getBoundingClientRect();
+        if (rect.width > 5 && rect.height > 5) {
+          svgs.push({
+            idx,
+            x: rect.x, y: rect.y, width: rect.width, height: rect.height
+          });
+        }
+      });
+
+      // 2. Containers vs Dividers
+      const containers = [];
+      const dividers = [];
+      const allDivs = Array.from(document.querySelectorAll("div, section, article"));
+      for (const el of allDivs) {
+        if (el === rootDiv || isInsideSvg(el)) continue;
+        const rect = el.getBoundingClientRect();
+        if (rect.width < 10 || rect.height < 2) continue;
+        if (rect.width >= 1270 && rect.height >= 710) continue;
+
+        const style = window.getComputedStyle(el);
+        const bg = parseRgba(style.backgroundColor);
+        const borderTopW = parseFloat(style.borderTopWidth) || 0;
+        const borderBottomW = parseFloat(style.borderBottomWidth) || 0;
+        const borderLeftW = parseFloat(style.borderLeftWidth) || 0;
+        const borderRightW = parseFloat(style.borderRightWidth) || 0;
+        const borderCol = parseRgba(style.borderTopColor);
+        const hasBg = bg && bg.a > 0.05;
+
+        if (!hasBg && borderTopW > 0 && borderLeftW === 0 && borderRightW === 0 && borderBottomW === 0) {
+          dividers.push({
+            x: rect.x,
+            y: rect.y,
+            width: rect.width,
+            color: borderCol || { r: 255, g: 255, b: 255, a: 0.1 }
+          });
+          continue;
+        }
+
+        const isFourSided = (borderTopW > 0 && (borderLeftW > 0 || borderRightW > 0 || borderBottomW > 0));
+        const hasBorder = isFourSided && borderCol && borderCol.a > 0.05 && style.borderTopStyle !== "none";
+
+        if (hasBg || hasBorder) {
+          containers.push({
+            x: rect.x,
+            y: rect.y,
+            width: rect.width,
+            height: rect.height,
+            bg: hasBg ? bg : null,
+            border: hasBorder ? { width: Math.max(borderTopW, borderLeftW), color: borderCol } : null,
+            borderRadius: parseFloat(style.borderRadius) || 0
+          });
+        }
+      }
+
+      // 3. Text Elements
+      const blockTags = ["H1","H2","H3","H4","H5","H6","P","LI"];
+      const inlineTags = ["SPAN","STRONG","B","EM","I","SMALL","LABEL","A","BUTTON"];
+
+      const textCandidates = [];
+      const allEls = Array.from(document.querySelectorAll("*"));
+      for (const el of allEls) {
+        if (isInsideSvg(el)) continue;
+        const tag = el.tagName;
+        if (blockTags.includes(tag)) {
+          textCandidates.push(el);
+        } else if (inlineTags.includes(tag)) {
+          if (!el.closest("h1, h2, h3, h4, h5, h6, p, li")) {
+            textCandidates.push(el);
+          }
+        } else if (tag === "DIV") {
+          const hasBlockChildren = Array.from(el.children).some(c => blockTags.includes(c.tagName) || c.tagName === "DIV");
+          const hasDirectText = Array.from(el.childNodes).some(n => n.nodeType === Node.TEXT_NODE && n.textContent.trim().length > 0);
+          if (!hasBlockChildren && hasDirectText && !el.closest("h1, h2, h3, h4, h5, h6, p, li") && !el.querySelector("svg")) {
+            textCandidates.push(el);
+          }
+        }
+      }
+
+      const textElements = [];
+      for (const el of textCandidates) {
+        const fullText = (el.innerText || "").trim();
+        if (!fullText) continue;
+
+        let rect = el.getBoundingClientRect();
+        try {
+          const range = document.createRange();
+          for (const node of el.childNodes) {
+            if (node.nodeType === Node.TEXT_NODE && node.textContent.trim()) {
+              range.selectNode(node);
+              const r = range.getBoundingClientRect();
+              if (r.width > 2 && r.height > 2) {
+                rect = r;
+                break;
+              }
+            }
+          }
+        } catch (e) {}
+
+        if (rect.width < 5 || rect.height < 5) continue;
+
+        const style = window.getComputedStyle(el);
+        const color = parseRgba(style.color) || { r: 255, g: 255, b: 255, a: 1 };
+        const weight = parseInt(style.fontWeight, 10) || (style.fontWeight === "bold" ? 700 : 400);
+
+        textElements.push({
+          tag: el.tagName,
+          text: fullText,
+          x: rect.x,
+          y: rect.y,
+          width: rect.width,
+          height: rect.height,
+          fontSize: parseFloat(style.fontSize) || 16,
+          fontWeight: weight,
+          color: color,
+          textAlign: style.textAlign
+        });
+      }
+
+      return {
+        bgColor: parseGradientOrColor(rootStyle),
+        containers,
+        dividers,
+        textElements,
+        svgs
+      };
+    });
+
+    // Capture exact background screenshot preserving root container's background & gradients
+    let bgBase64 = null;
+    try {
+      await page.evaluate(() => {
+        const root = document.body.firstElementChild || document.body;
+        Array.from(root.children).forEach(child => {
+          if (child.tagName !== "STYLE" && child.tagName !== "SCRIPT") {
+            child.dataset.origVisibility = child.style.visibility || "";
+            child.style.visibility = "hidden";
+          }
+        });
+      });
+      bgBase64 = await page.screenshot({ encoding: "base64" });
+      await page.evaluate(() => {
+        const root = document.body.firstElementChild || document.body;
+        Array.from(root.children).forEach(child => {
+          if (child.dataset && child.dataset.origVisibility !== undefined) {
+            child.style.visibility = child.dataset.origVisibility;
+            delete child.dataset.origVisibility;
+          }
+        });
+      });
+    } catch (e) {}
+
+    data.bgBase64 = bgBase64;
+
+    await page.evaluate(() => {
+      document.documentElement.style.setProperty("background", "transparent", "important");
+      document.body.style.setProperty("background", "transparent", "important");
+      document.querySelectorAll("*").forEach(el => {
+        if (el.tagName !== "svg" && !el.closest("svg")) {
+          el.style.setProperty("background", "transparent", "important");
+        }
+      });
+    });
+
+    const svgEls = await page.$$("svg");
+    for (let i = 0; i < svgEls.length; i++) {
+      try {
+        const b64 = await svgEls[i].screenshot({ omitBackground: true, encoding: "base64" });
+        if (data.svgs[i]) {
+          data.svgs[i].base64 = b64;
+        }
+      } catch (e) {}
+    }
+
+    results.push(data);
+    await page.close();
+  }
+
+  console.log(JSON.stringify(results));
+  await browser.close();
+})();
+"""
+
+
+def _extract_slide_elements_via_chromium(slide_items: List[Dict[str, Any]]) -> Optional[List[Dict[str, Any]]]:
+    """
+    Extracts layout geometry, containers, text boxes, and SVGs from HTML slides via headless Chromium.
+    Returns structured list of extracted slide dictionaries, or None if unavailable.
+    """
+    browser_bin = shutil.which("chromium") or shutil.which("chromium-browser") or shutil.which("google-chrome")
+    node_bin = shutil.which("node")
+    if not browser_bin or not node_bin:
+        return None
+
+    cleaned_items = []
+    for it in slide_items:
+        cleaned_items.append({
+            "key": it.get("key", ""),
+            "title": it.get("title", ""),
+            "html": _sanitize_slide_html(it.get("html", "")),
+            "notes": it.get("notes", "")
+        })
+
+    env = os.environ.copy()
+    env["NODE_PATH"] = "/usr/lib/node_modules"
+    try:
+        proc = subprocess.run(
+            [node_bin, "-e", NODE_DOM_EXTRACTION_SCRIPT],
+            input=json.dumps(cleaned_items).encode("utf-8"),
+            capture_output=True,
+            timeout=45,
+            env=env
+        )
+        if proc.returncode == 0 and proc.stdout:
+            return json.loads(proc.stdout.decode("utf-8"))
+    except Exception as e:
+        print(f"Chromium layout extraction fallback due to: {e}")
+        return None
+    return None
+
+
+def _parse_rgb_tuple(c_str: Optional[str], fallback=(15, 23, 42)):
+    if not c_str:
+        return fallback
+    if c_str.startswith("#"):
+        h = c_str.lstrip("#")
+        if len(h) == 3:
+            h = "".join([x * 2 for x in h])
+        try:
+            return (int(h[:2], 16), int(h[2:4], 16), int(h[4:6], 16))
+        except Exception:
+            return fallback
+    m = re.findall(r"\d+", c_str)
+    if len(m) >= 3:
+        return (int(m[0]), int(m[1]), int(m[2]))
+    return fallback
+
+
+def compile_to_pptx(artifact: SessionArtifact) -> io.BytesIO:
+    """
+    Compile Presentation artifact into 100% compliant 16:9 Microsoft PowerPoint (.pptx)
+    compatible with Apple Keynote, Google Slides, and MS Office.
+
+    Guarantees 100% editable text boxes, native styled shapes/cards, and crisp transparent
+    vector graphics (SVGs), with zero rasterized slide screenshot flattening.
+    """
+    prs = Presentation()
+    # 16:9 widescreen layout (13.333333 x 7.5 inches = 12192000 x 6858000 EMUs)
+    prs.slide_width = Inches(13.333333)
+    prs.slide_height = Inches(7.5)
+
+    # Explicitly set screen16x9 type in PresentationML
+    if hasattr(prs._element, "sldSz") and prs._element.sldSz is not None:
+        prs._element.sldSz.attrib['type'] = 'screen16x9'
+
+    blank_layout = prs.slide_layouts[6]
+    blocks = sorted(artifact.blocks, key=lambda b: b.order_index) if artifact.blocks else []
+
+    slide_items = []
+    if not blocks:
+        slide_items.append({
+            "key": "slide_1",
+            "title": artifact.title or "Presentation Deck",
+            "html": f"<div style='width: 100%; height: 100%; display: flex; flex-direction: column; justify-content: center; align-items: center; background: #0f172a; color: #fff; font-family: sans-serif;'><h1>{artifact.title or 'Presentation Deck'}</h1><p>Generative Presentation</p></div>",
+            "notes": ""
+        })
+    else:
+        for idx, b in enumerate(blocks):
+            raw = (b.content or "").strip()
+            slide_html = ""
+            slide_notes = ""
+            slide_title = b.title or f"Slide {idx + 1}"
+
+            try:
+                data = json.loads(raw)
+                if isinstance(data, dict):
+                    slide_title = data.get("title") or b.title or f"Slide {idx + 1}"
+                    slide_notes = data.get("notes") or data.get("speaker_notes") or ""
+                    slide_html = data.get("html") or data.get("custom_html") or ""
+                    if not slide_html and ("cards" in data or "subtitle" in data or "bullets" in data):
+                        cards_html = ""
+                        for c in data.get("cards", []):
+                            val = c.get("value") or ""
+                            val_html = f"<div style='font-size: 28px; font-weight: 800; color: #38bdf8;'>{val}</div>" if val else ""
+                            cards_html += f"<div style='background: rgba(255,255,255,0.05); border: 1px solid rgba(255,255,255,0.1); border-radius: 12px; padding: 20px;'>{val_html}<div style='font-size: 16px; font-weight: 700; margin-bottom: 8px;'>{c.get('title', '')}</div><p style='font-size: 13px; color: #94a3b8;'>{c.get('content', '')}</p></div>"
+                        bullets_html = "".join(f"<li>{item}</li>" for item in data.get("bullets", []))
+                        slide_html = f"<div style='width: 100%; height: 100%; background: #0f172a; color: #fff; padding: 48px; box-sizing: border-box;'><h1 style='font-size: 38px; margin-bottom: 8px;'>{slide_title}</h1><p style='color: #94a3b8; font-size: 16px; margin-bottom: 24px;'>{data.get('subtitle', '')}</p><div style='display: grid; grid-template-columns: repeat(auto-fit, minmax(200px, 1fr)); gap: 16px;'>{cards_html}</div><ul>{bullets_html}</ul></div>"
+                elif isinstance(data, list):
+                    slide_html = f"<div><h1>{slide_title}</h1><ul>" + "".join(f"<li>{item}</li>" for item in data) + "</ul></div>"
+            except Exception:
+                if raw.startswith("<"):
+                    slide_html = raw
+                else:
+                    slide_html = f"<div><h1>{slide_title}</h1><p>{raw}</p></div>"
+
+            slide_items.append({
+                "key": b.block_key or f"slide_{idx + 1}",
+                "title": slide_title,
+                "html": slide_html,
+                "notes": slide_notes
+            })
+
+    # Attempt high-fidelity native DrawingML extraction via Chromium DOM layout engine
+    extracted_slides = _extract_slide_elements_via_chromium(slide_items)
+
+    scale_x = Inches(13.333333) / 1280.0
+    scale_y = Inches(7.5) / 720.0
+
+    if extracted_slides and len(extracted_slides) == len(slide_items):
+        for idx, r in enumerate(extracted_slides):
+            slide = prs.slides.add_slide(blank_layout)
+            bg_tuple = _parse_rgb_tuple(r.get("bgColor"))
+
+            # 1. Slide Background (Native solid color and crisp background graphic/layer)
+            fill = slide.background.fill
+            fill.solid()
+            fill.fore_color.rgb = RGBColor(*bg_tuple)
+
+            # Insert explicit background shape/image as base layer (shape index 0)
+            # This ensures Keynote and PowerPoint render the exact browser background color/gradient
+            bg_b64 = r.get("bgBase64")
+            if bg_b64:
+                import base64
+                bg_png = base64.b64decode(bg_b64)
+                slide.shapes.add_picture(io.BytesIO(bg_png), 0, 0, prs.slide_width, prs.slide_height)
+            else:
+                bg_shape = slide.shapes.add_shape(MSO_SHAPE.RECTANGLE, 0, 0, prs.slide_width, prs.slide_height)
+                bg_shape.fill.solid()
+                bg_shape.fill.fore_color.rgb = RGBColor(*bg_tuple)
+                bg_shape.line.fill.background()
+
+            # 2. Containers (Cards / Bento Boxes)
+            for c_box in r.get("containers", []):
+                c_left = int(c_box["x"] * scale_x)
+                c_top = int(c_box["y"] * scale_y)
+                c_w = int(c_box["width"] * scale_x)
+                c_h = int(c_box["height"] * scale_y)
+                st = MSO_SHAPE.ROUNDED_RECTANGLE if c_box.get("borderRadius", 0) > 4 else MSO_SHAPE.RECTANGLE
+                shape = slide.shapes.add_shape(st, c_left, c_top, c_w, c_h)
+
+                bg = c_box.get("bg")
+                if bg:
+                    alpha = bg.get("a", 1.0)
+                    br = int(bg_tuple[0] * (1 - alpha) + bg["r"] * alpha)
+                    bg_g = int(bg_tuple[1] * (1 - alpha) + bg["g"] * alpha)
+                    bb = int(bg_tuple[2] * (1 - alpha) + bg["b"] * alpha)
+                    shape.fill.solid()
+                    shape.fill.fore_color.rgb = RGBColor(br, bg_g, bb)
+                else:
+                    shape.fill.background()
+
+                border = c_box.get("border")
+                if border and border.get("color"):
+                    b_col = border["color"]
+                    shape.line.color.rgb = RGBColor(b_col["r"], b_col["g"], b_col["b"])
+                    shape.line.width = Pt(max(1, border.get("width", 1)))
+                else:
+                    shape.line.fill.background()
+
+            # 3. Dividers (Subtle 1px divider lines)
+            for d_line in r.get("dividers", []):
+                d_left = int(d_line["x"] * scale_x)
+                d_top = int(d_line["y"] * scale_y)
+                d_w = int(d_line["width"] * scale_x)
+                line_shape = slide.shapes.add_shape(MSO_SHAPE.RECTANGLE, d_left, d_top, d_w, Pt(1))
+                d_col = d_line.get("color", {"r": 100, "g": 116, "b": 139})
+                line_shape.fill.solid()
+                line_shape.fill.fore_color.rgb = RGBColor(d_col["r"], d_col["g"], d_col["b"])
+                line_shape.line.fill.background()
+
+            # 4. SVGs & Vector Graphics (100% Transparent Crisply Isolated Pictures)
+            import base64
+            for svg in r.get("svgs", []):
+                b64 = svg.get("base64")
+                if b64:
+                    raw_png = base64.b64decode(b64)
+                    s_left = int(svg["x"] * scale_x)
+                    s_top = int(svg["y"] * scale_y)
+                    s_w = int(svg["width"] * scale_x)
+                    s_h = int(svg["height"] * scale_y)
+                    slide.shapes.add_picture(io.BytesIO(raw_png), s_left, s_top, s_w, s_h)
+
+            # 5. Text Boxes (100% Native Editable Text with Word Wrap and Accurate Typography)
+            for t in r.get("textElements", []):
+                t_left = int(t["x"] * scale_x)
+                t_top = int(t["y"] * scale_y)
+                t_w = int(t["width"] * scale_x) + Inches(0.2)
+                t_h = int(t["height"] * scale_y) + Inches(0.08)
+
+                box = slide.shapes.add_textbox(t_left, t_top, t_w, t_h)
+                tf = box.text_frame
+                tf.word_wrap = True
+                tf.margin_left = Inches(0)
+                tf.margin_right = Inches(0)
+                tf.margin_top = Inches(0)
+                tf.margin_bottom = Inches(0)
+
+                p = tf.paragraphs[0]
+                p.text = t["text"]
+                f_pt = max(8.5, t["fontSize"] * 0.75)
+                p.font.size = Pt(f_pt)
+                p.font.bold = t["fontWeight"] >= 600
+
+                col = t.get("color", {"r": 255, "g": 255, "b": 255})
+                p.font.color.rgb = RGBColor(col["r"], col["g"], col["b"])
+
+                align = t.get("textAlign", "left")
+                if align == "center":
+                    p.alignment = PP_ALIGN.CENTER
+                elif align == "right":
+                    p.alignment = PP_ALIGN.RIGHT
+
+            # 6. Speaker Notes
+            notes = slide_items[idx].get("notes", "")
+            if notes and notes.strip():
+                slide.notes_slide.notes_text_frame.text = notes.strip()
+    else:
+        # Graceful fallback to pure Python DrawingML DOM compiler
+        for item in slide_items:
+            slide = prs.slides.add_slide(blank_layout)
+            _compile_generative_slide_dom(
+                slide,
+                slide_html=item.get("html", ""),
+                default_title=item.get("title", "Slide Title"),
+                notes_text=item.get("notes", "")
+            )
+
+    _ensure_notes_master_registered(prs)
+
+    buf = io.BytesIO()
+    prs.save(buf)
+    buf.seek(0)
+    return buf
+
